DTSPB-1184-update tests
